+++ Model/Products/Product.cs
@@ -9,14 +9,14 @@ namespace CheckList
     public class Product
     {
         private string  _name, _description;
-        private double _price;
+        private float _price;
         private int _stock, _id;
 
         public Product()
         {
         }
 
-        public Product(int id, string name, string description, double price, int stock)
+        public Product(int id, string name, string description, float price, int stock)
         {
             _id = id;
             _name = name;
@@ -27,7 +27,7 @@ namespace CheckList
         public int Id { get => _id; set => _id = value; }
         public string Name { get => _name; set => _name = value; }
         public string Description { get => _description; set => _description = value; }
-        public double Price { get => _price; set => _price = value; }
+        public float Price { get => _price; set => _price = value; }
         public int Stock { get => _stock; set => _stock = value; }
     }
 }
